# scratch_swagger_new.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "StreamVault",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/WatchHistory":  {
                                            "post":  {
                                                         "tags":  [
                                                                      "WatchHistory"
                                                                  ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/WatchHistoryCreateRequest"
                                                                                                                                 }
                                                                                                                  },
                                                                                             "text/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/WatchHistoryCreateRequest"
                                                                                                                          }
                                                                                                           },
                                                                                             "application/*+json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/WatchHistoryCreateRequest"
                                                                                                                                   }
                                                                                                                    }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     }
                                        },
                  "/api/WatchHistory/paged":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "WatchHistory"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "page",
                                                                                     "in":  "query",
                                                                                     "schema":  {
                                                                                                    "type":  "integer",
                                                                                                    "format":  "int32",
                                                                                                    "default":  1
                                                                                                }
                                                                                 },
                                                                                 {
                                                                                     "name":  "size",
                                                                                     "in":  "query",
                                                                                     "schema":  {
                                                                                                    "type":  "integer",
                                                                                                    "format":  "int32",
                                                                                                    "default":  10
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK"
                                                                                        }
                                                                            }
                                                          }
                                              },
                  "/api/Watchlist":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "Watchlist"
                                                               ],
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/WatchlistCreateRequest"
                                                                                                                              }
                                                                                                               },
                                                                                          "text/json":  {
                                                                                                            "schema":  {
                                                                                                                           "$ref":  "#/components/schemas/WatchlistCreateRequest"
                                                                                                                       }
                                                                                                        },
                                                                                          "application/*+json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/WatchlistCreateRequest"
                                                                                                                                }
                                                                                                                 }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK"
                                                                                }
                                                                    }
                                                  }
                                     },
                  "/api/Watchlist/{watchlistGuid}":  {
                                                         "delete":  {
                                                                        "tags":  [
                                                                                     "Watchlist"
                                                                                 ],
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "watchlistGuid",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  {
                                                                                                              "type":  "string"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "OK"
                                                                                                  }
                                                                                      }
                                                                    }
                                                     },
                  "/api/Watchlist/user/{userGuid}":  {
                                                         "get":  {
                                                                     "tags":  [
                                                                                  "Watchlist"
                                                                              ],
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "userGuid",
                                                                                            "in":  "path",
                                                                                            "required":  true,
                                                                                            "schema":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "OK"
                                                                                               }
                                                                                   }
                                                                 }
                                                     }
              },
    "components":  {
                       "schemas":  {
...
                                       "WatchHistoryCreateRequest":  {
                                                                         "type":  "object",
                                                                         "properties":  {
                                                                                            "userGuid":  {
                                                                                                             "type":  "string",
                                                                                                             "nullable":  true
                                                                                                         },
                                                                                            "movieGuid":  {
                                                                                                              "type":  "string",
                                                                                                              "nullable":  true
                                                                                                          },
                                                                                            "watchedDate":  {
                                                                                                                "type":  "string",
                                                                                                                "format":  "date-time"
                                                                                                            },
                                                                                            "watchMinutes":  {
                                                                                                                 "type":  "integer",
                                                                                                                 "format":  "int32"
                                                                                                             },
                                                                                            "deviceType":  {
                                                                                                               "type":  "string",
                                                                                                               "nullable":  true
                                                                                                           }
                                                                                        },
                                                                         "additionalProperties":  false
                                                                     },
                                       "WatchlistCreateRequest":  {
                                                                      "type":  "object",
                                                                      "properties":  {
                                                                                         "movieGuid":  {
                                                                                                           "type":  "string",
                                                                                                           "nullable":  true
                                                                                                       }
                                                                                     },
                                                                      "additionalProperties":  false
                                                                  }
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "apiKey",
                                                              "description":  "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter \u0027Bearer\u0027 [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
                                                              "name":  "Authorization",
                                                              "in":  "header"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}